time: Add epoch time implementation
This patch provides an interface to manage the epoch time based on RTC
and the hardware clock.

With the help of RTC, we can keep the epoch time even when the system
is off. When the system is up and running, we can provide a fine
resolution time unit, i.e. microseconds, based on hardware clock (though
the resolution is still depends on the clock frequency of hardware
clock).

Initially, it will read the current epoch time from RTC and keep current
hardware clock cycles as the baseline. When the epoch_time_get() is
called, the elapsed time will be calculated based on the difference of
current hardware clock cycles and the baseline. If RTC is not available,
it will assume current time is 00:00:00 UTC on 1 January 1970.

A new option "CONFIG_EPOCH_TIME_RTC_WRITE_THROUGH" is introduced. By
enabling this option, if current time is changed by epoch_time_set(),
the change will also be applied to RTC. So the time will be on the track
even if system is off.

Signed-off-by: Aska Wu <aska.wu@linaro.org>